chore: change NodesUsed, PVCCount and JobCount to int
The variables NodesUsed, PVCCount and JobCount in the status section
of the Cluster type were fixed to int32, this was triggering the gosec
linter due to casting from int to int32 when it's not needed, all the
operations are coming and process as int.

+++ internal/controller/cluster_status.go
@@ -250,7 +250,7 @@ func (r *ClusterReconciler) updateResourceStatus(
 	)
 
 	// Count jobs
-	newJobs := int32(len(resources.jobs.Items))
+	newJobs := len(resources.jobs.Items)
 	cluster.Status.JobCount = newJobs
 
 	cluster.Status.Topology = getPodsTopology(
@@ -784,7 +784,7 @@ func getPodsTopology(
 		}
 	}
 
-	return apiv1.Topology{SuccessfullyExtracted: true, Instances: data, NodesUsed: int32(len(nodesMap))}
+	return apiv1.Topology{SuccessfullyExtracted: true, Instances: data, NodesUsed: len(nodesMap)}
 }
 
 // isWALSpaceAvailableOnPod check if a Pod terminated because it has no
